GafferAstro : Conform plug creation
Switches to dict style assignment over `addChild`. The main merit
being that it ensures you are using the plug you are expecting to
later on, when referring to one by name. Though this is largely
true either way, it can be more obvious when you assign to the same
child name multiple times in this form.