Spreadsheet : Allow addColumn serialisation to be skipped
Presently, the standard Gaffer serialisation of Spreadsheets
outputs an `addColumn` call for each column in the sheet.
If a spreadsheet column is added during initialisation of a
custom node, it will be duplicated by this serialisation.

As a work around, until a proper Gaffer mechanism is put in place,
this allows custom nodes to ask specific columns not to be
serialised via the `serialiser:spreadsheet:skipColumn` metadata
on a default row cell.